Little bits of nothing.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@8121 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Behaviors/ResourceOverride.java
@@ -60,14 +60,14 @@ public class ResourceOverride extends StdBehavior
 				for(int i=0;i<RawMaterial.MATERIAL_DESCS.length;i++)
 				{
 					if(RawMaterial.MATERIAL_DESCS[i].equalsIgnoreCase(which))
-					{ code=RawMaterial.CODES.COMPOSE_RESOURCES(i).get(0); break;}
+					{ code=RawMaterial.CODES.COMPOSE_RESOURCES(i).get(0).intValue(); break;}
 				}
 			if(code<0) code = RawMaterial.CODES.FIND_StartsWith(which);
 			if(code<0)
 				for(int i=0;i<RawMaterial.MATERIAL_DESCS.length;i++)
 				{
 					if(RawMaterial.MATERIAL_DESCS[i].startsWith(which))
-					{ code=RawMaterial.CODES.COMPOSE_RESOURCES(i).get(0); break;}
+					{ code=RawMaterial.CODES.COMPOSE_RESOURCES(i).get(0).intValue(); break;}
 				}
 			if(code>=0)
 			{
